refactor(MapListnerHandlers): delegate to three subfiles
